done location-view sub-component
