+++ src/Template/Element/Form/locations.scss
@@ -21,5 +21,21 @@
             border-color: $white;
             color: $white;
         }
+
+        .buttons {
+            justify-content: space-between;
+            margin-right: 1rem;
+
+            button {
+                text-transform: uppercase;
+                background-color: $defaultBackground;
+                border-color: $white;
+                color: $white;
+            }
+
+            .remove {
+                border: none;
+            }
+        }
     }
 }
