+++ src/Template/Layout/js/app/components/property-view/property-view.js
@@ -21,7 +21,7 @@ export default {
         MapView: () => import(/* webpackChunkName: "map-view" */'app/components/map-view'),
         DateRangesView: () => import(/* webpackChunkName: "date-ranges-view" */'app/components/date-ranges-view/date-ranges-view'),
         History: () => import(/* webpackChunkName: "history" */'app/components/history'),
-        LocationsView: () => import(/* webpackChunkName: "locations-view" */'app/components/locations-view'),
+        LocationsView: () => import(/* webpackChunkName: "locations-view" */'app/components/locations-view/locations-view'),
     },
 
     props: {
